enh: reply to yjs sync step 1 as the server
The yjs sync protocol has three types of messages:
* Sync step 1 (Question)
* Sync step 2 (Answer)
* Update

The different parties will exchange questions and answers to sync.
Once all parties are synced they exchange updates containing the latest changes.

We used to relay all questions to all parties
which resulted in a lot of unneccessary traffic and db entries.

Instead only store the answers and updates on the server.
Whenever any question comes in reply with the log of all the other messages.
This way questions are only answered once (by the server)
and we can still stick to a "stupid" server
that does not need to understand state vectors and so on.

Signed-off-by: Max <max@nextcloud.com>